% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_missing_flows}
\alias{add_missing_flows}
\title{Appends missing rows to flow data}
\usage{
add_missing_flows(flows, regions, agegroups, years)
}
\arguments{
\item{flows}{data.table with flows}

\item{regions}{character, unique regions}

\item{agegroups}{character, unique agegroups}

\item{years}{numeric, unique years}
}
\value{
flow data.table with missing rows
}
\description{
Flow data obtained from the federal statistical office only has
rows if flows are > 0. add_missing_flows adds rows that are
missing and sets flows for those rows to 0.
}
\author{
Konstantin Hoffie
}
