% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cut_qs}
\alias{cut_qs}
\title{Cuts a numeric vector based on quantiles. Either wih evenly sized
quantiles or with hand chosen ones.}
\usage{
cut_qs(
  x,
  num_quantiles = NULL,
  probs = cut_probs(),
  labels = NULL,
  lab = FALSE
)
}
\arguments{
\item{x}{numeric vector}

\item{num_quantiles}{integer, optional}

\item{probs}{vector of quantiles}

\item{labels}{character, optional vector to label resulting vector}

\item{lab}{logical, if FALSE vector has integers instead of labels}
}
\value{
numeric, cutted vector
}
\description{
Cuts a numeric vector based on quantiles. Either wih evenly sized
quantiles or with hand chosen ones.
}
\author{
Konstantin Hoffie
}
