% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggsavew}
\alias{ggsavew}
\title{Simple wrapper around ggplot2::ggsave to easily change plot size.}
\usage{
ggsavew(w = 8, ...)
}
\arguments{
\item{w}{width}

\item{...}{arguments passed to ggplot2}
}
\description{
Simple wrapper around ggplot2::ggsave to easily change plot size.
}
\author{
Konstantin Hoffie
}
